+++ hub-client/src/assets/icons.ts
@@ -398,7 +398,7 @@ const icons: { [key: string]: string } = {
         </svg>
     `,
 	promote: `
-        <svg class="h-8 w-8 text-gray-500"  fill="none" viewBox="0 0 24 24" stroke="currentColor">
+        <svg class="h-8 w-8"  fill="none" viewBox="0 0 24 24" stroke="currentColor">
         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 11l3-3m0 0l3 3m-3-3v8m0-13a9 9 0 110 18 9 9 0 010-18z"/>
         </svg>
    `,