+++ hub-client/src/components/elements/H1.vue
@@ -1,5 +1,5 @@
 <template>
-	<h1 v-tw-class="'font-body font-bold ~text-h1-min/h1-max'">
+	<h1 v-tw-class="'font-headings font-semibold ~text-h1-min/h1-max'">
 		<slot></slot>
 	</h1>
 </template>