+++ hub-client/src/components/elements/H2.vue
@@ -1,5 +1,5 @@
 <template>
-	<h2 v-tw-class="'font-body font-bold ~text-h2-min/h2-max'">
+	<h2 v-tw-class="'font-headings font-semibold ~text-h2-min/h2-max'">
 		<slot></slot>
 	</h2>
 </template>