+++ hub-client/src/pages/ManageUsers.vue
@@ -8,8 +8,8 @@
 				<span class="font-semibold uppercase">{{ t('admin.title_administrator') }}</span>
 				<hr class="h-[2px] grow bg-on-surface-dim" />
 			</div>
-			<div class="flex h-full items-center pl-12 md:pl-0">
-				<H3 class="font-body font-bold text-on-surface">{{ t('menu.admin_tools_users') }}</H3>
+			<div class="flex h-full items-center" :class="isMobile ? 'pl-12' : 'pl-0'">
+				<H3 class="font-headings font-semibold text-on-surface">{{ t('menu.admin_tools_users') }}</H3>
 			</div>
 		</template>
 
@@ -25,16 +25,16 @@
 			<!---List all users accounts -->
 			<FilteredList :items="hubUsers" :filterKey="['displayname']" sortby="displayname" :placeholder="$t('rooms.filter')">
 				<template #item="{ item }">
-					<div class="flex w-full justify-between gap-8" :title="item.room_id">
-						<div class="flex w-full items-center gap-4">
+					<div class="box-border flex w-full justify-between gap-4 md:gap-8" :title="item.room_id">
+						<div class="flex min-w-0 flex-1 items-center gap-4">
 							<Avatar :user="item.name" :override-avatar-url="item.avatar_url" />
-							<p class="truncate font-semibold">{{ item.displayname }}</p>
-							<p class="hidden truncate pr-1 italic text-on-surface-dim md:inline">{{ item.name }}</p>
-							<span v-if="item.admin" class="relative items-center rounded-md bg-accent-red px-1 font-medium text-on-accent-red ~text-label-min/label-max">Hub Administrator</span>
+							<p class="min-w-0 truncate font-semibold">{{ item.displayname }}</p>
+							<p class="line-clamp-1 hidden min-w-0 pr-1 italic text-on-surface-dim md:inline">{{ item.name }}</p>
+							<span v-if="item.admin" class="relative line-clamp-1 w-fit items-center rounded-md bg-accent-red px-1 font-medium text-on-accent-red ~text-label-min/label-max">Hub Administrator</span>
 						</div>
 						<div class="flex w-fit gap-4">
 							<div class="flex items-center gap-2">
-								<Icon type="edit" class="hover:text-accent-primary" @click="selectUser(item.name, item.displayname, item.avatar_url)" />
+								<Icon type="edit" class="hover:cursor-pointer hover:text-accent-primary" @click="selectUser(item.name, item.displayname, item.avatar_url)" />
 							</div>
 						</div>
 					</div>
@@ -47,14 +47,17 @@
 <script setup lang="ts">
 	import Avatar from '@/components/ui/Avatar.vue';
 
-	import { onMounted, ref } from 'vue';
+	import { computed, onMounted, ref } from 'vue';
 	import { TUserAccount } from '@/model/users/TUser';
 	import { useUser } from '@/logic/store/user';
 	import { ManagementUtils } from '@/model/hubmanagement/utility/managementutils';
 
 	import { useI18n } from 'vue-i18n';
+	import { useSettings } from '@/logic/store/settings';
 
 	const { t } = useI18n();
+	const settings = useSettings();
+	const isMobile = computed(() => settings.isMobileState);
 
 	// Store
 	const user = useUser();